requirements.txt
# BSE Corporate Announcement Monitor — Python dependencies
bse>=3.2.0          # Unofficial BSE India API (BseIndiaApi)
schedule>=1.2.0     # Job scheduler for polling intervals
requests>=2.31.0    # HTTP requests (PDF downloads)
pdfplumber>=0.11.0  # PDF text extraction
google-generativeai>=0.8.0  # Google Gemini API for summarization
# callmebot uses plain HTTP (requests already listed above) — no extra package needed
python-dotenv>=1.0.0  # .env file support
pyyaml>=6.0         # config.yaml parsing
